Changed gameLoop order
Changed order of operations executed within gameLoop -- after creating a simple test game state, it appeared that rendering was not working properly. Changing the order fixed this. As stated before, a new test game state was added as well. The package naming was changed too, to reflect the new name of developer.